// apis/share-api.ts
export const GetShareDetailDetailEnum = {
    NUMBER_1: 1
} as const;
export type GetShareDetailDetailEnum = typeof GetShareDetailDetailEnum[keyof typeof GetShareDetailDetailEnum];
// ...
export const ListSharesOrderByEnum = {
    CreateTime: 'createTime',
    ExpireTime: 'expireTime',
    Name: 'name',
    CreatorId: 'creatorId',
    Empty: ''
} as const;
export type ListSharesOrderByEnum = typeof ListSharesOrderByEnum[keyof typeof ListSharesOrderByEnum];
export const ListSharesOrderByTypeEnum = {
    Asc: 'asc',
    Desc: 'desc'
} as const;
export type ListSharesOrderByTypeEnum = typeof ListSharesOrderByTypeEnum[keyof typeof ListSharesOrderByTypeEnum];
export const ListSharesWithFileInfoEnum = {
    NUMBER_0: 0,
    NUMBER_1: 1
} as const;
export type ListSharesWithFileInfoEnum = typeof ListSharesWithFileInfoEnum[keyof typeof ListSharesWithFileInfoEnum];